Moving shared Error and Testing objects into common.py
